# lib/api-zod/src/generated/types/favoriteTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from "./track";

export type FavoriteTrack = Track & {
  createdAt: string;
};