# lib/api-zod/src/generated/types/playerState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerState {
  currentVideoId?: string | null;
  isPlaying: boolean;
  updatedBy?: string | null;
  updatedAt: string;
  streamUrl?: string | null;
}