# lib/api-zod/src/generated/types/playlistTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from "./track";

export type PlaylistTrack = Track & {
  addedBy: string;
  addedAt: string;
  isFavorite: boolean;
};