# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  videoId: string;
  title: string;
  artist: string;
  duration: string;
  thumbnail?: string | null;
  streamUrl: string;
}