# lib/api-zod/src/generated/types/updatePlayerStateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePlayerStateRequest {
  currentVideoId?: string | null;
  isPlaying: boolean;
  updatedBy?: string;
}